+++ src/com/SGA/Athlete.java
@@ -12,8 +12,20 @@ public class Athlete {
     }
 
     private SexCategory sex;
-
-    private HashMap<CompetitionType, Double> competitionResults = new HashMap<>();
+    /*
+    private double running60;
+    private double running200;
+    private double running800;
+    private double running1500;
+    private double running3000;
+    private double runningHurdles60;
+    private double jumpingLong;
+    private double jumpingTriple;
+    private double jumpingHigh;
+    private double jumpingPole;
+    private double throwingShot;
+    */
+    private HashMap<CompetitionType, Double> competitionResults;
 
 
     // Getters and setters
@@ -119,13 +131,4 @@ public class Athlete {
         }
         return result;
     }
-
-    public void printAthleteInformation(){
-        System.out.println("First Name : " + this.firstName);
-        System.out.println("Last Name  : " + this.lastName );
-        System.out.println("Sex        : " + this.sex);
-        System.out.println("Results    : " + this.competitionResults);
-
-    }
-
 }
