+++ src/com/SGA/Athletes.java
@@ -1,0 +1,7 @@
+package com.SGA;
+
+import java.util.Vector;
+
+public class Athletes {
+    Vector<Athlete> athletes;
+}
