+++ src/com/SGA/PoleJumping.java
@@ -1,8 +1,9 @@
 package com.SGA;
 
-public class JumpingHigh extends Competition{
+public class PoleJumping extends Competition {
     @Override
     public TimeSlot getNext() {
+        System.out.println("PoleJumping");
         return super.getNext();
     }
 }
