--- src/com/SGA/Station.java
@@ -1,11 +1,0 @@
-package com.SGA;
-
-public enum Station {
-    RUNNINGCIRCLE,
-    SPRINTLINE,
-    LONGTRIPLEJUMP,
-    HIGHJUMP,
-    POLEVAULT,
-    SHOTTHROWING,
-    AWARDCEREMONYAREA
-}
