+++ src/com/SGA/TimeSlot.java
@@ -1,0 +1,6 @@
+package com.SGA;
+
+public class TimeSlot {
+    public int startTime;
+    public int endTime;
+}
